Move Truffle::Ropes.debug_print to Truffle::Debug.print
* It is not specific to ropes and is a general debug utility.
* Remove empty spec.
