--- spec/truffle/ropes/debug_print_spec.rb
@@ -1,15 +1,0 @@
-# Copyright (c) 2015, 2017 Oracle and/or its affiliates. All rights reserved. This
-# code is released under a tri EPL/GPL/LGPL license. You can use it,
-# redistribute it and/or modify it under the terms of the:
-# 
-# Eclipse Public License version 1.0
-# GNU General Public License version 2
-# GNU Lesser General Public License version 2.1
-
-require_relative '../../ruby/spec_helper'
-
-describe "Truffle::Ropes.debug_print" do
-  
-  it "needs to be reviewed for spec completeness"
-
-end
